requirements.txt
# Core dependencies
streamlit==1.29.0
python-dotenv==1.0.0
flask==3.0.0
flask-cors==4.0.0

# Web scraping
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3
selenium==4.15.0

# Data processing
pandas==2.1.4
numpy==1.24.3

# NLP and Embeddings
sentence-transformers==2.2.2
transformers==4.36.0
torch==2.1.0

# Vector database
chromadb==0.4.18
# faiss-cpu==1.7.4  # Alternative option

# LLM Integration
openai==1.6.1
langchain==0.1.0
langchain-community==0.0.10

# Evaluation
scikit-learn==1.3.2

# Utilities
pyyaml==6.0.1
tqdm==4.66.1
